+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - **Breaking:** Return early with an error when any affected file is uncommitted, unless `--force`d
+- Add better support for glob imports
 
 ## [0.5.0] - 2025-07-13